css 적용

+++ src/components/BangWeet.js
@@ -1,14 +1,21 @@
 import React, {useState} from "react";
 import {dbService, storageService} from "../fbase";
+import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
+import {faTrash, faPencilAlt} from "@fortawesome/free-solid-svg-icons";
+
 
 const BangWeet = ({bangWeetObj, isOwner}) => {
     const [editing, setEditing] = useState(false);
     const [newBangWeet, setNewBangWeet] = useState(bangWeetObj.text);
+
     const onDeleteClick = async () => {
         const ok = window.confirm("삭제하시겠습니까?");
         if (ok) {
-            await storageService.refFromURL(bangWeetObj.attachmentUrl).delete();
+            console.log(bangWeetObj)
             await dbService.doc(`bangWeets/${bangWeetObj.id}`).delete();
+            if (bangWeetObj.attachmentUrl != "") {
+                await storageService.refFromURL(bangWeetObj.attachmentUrl).delete();
+            }
         }
     }
     const toggleEditting = () => setEditing((prev) => !prev);
@@ -25,34 +32,34 @@ const BangWeet = ({bangWeetObj, isOwner}) => {
     }
 
     return (
-        <div>
+        <div className="bangWeet">
             {editing ?
                 <>
                     {isOwner && (
                         <>
-                            <form onSubmit={onSubmit}>
+                            <form onSubmit={onSubmit} className="container bangWeetEdit">
                                 <input type="text" placeholder="edit your bangWeet" onChange={onChange}
                                        value={newBangWeet}
                                        required/>
-                                <input type="submit" value="update bangWeet"/>
+                                <input type="submit" value="update bangWeet" className="formBtn"/>
                             </form>
-                            <button onClick={toggleEditting}>Cancel</button>
+                            <button onClick={toggleEditting} className="formBtn cancelBtn">Cancel</button>
                         </>
                     )}
                 </>
                 :
                 <>
                     <h4>{bangWeetObj.text}</h4>
-                    {bangWeetObj.attachmentUrl && (
-                        <>
-                            <img alt="attachment" src={bangWeetObj.attachmentUrl} width="100px" height="100px" />
-                        </>
-                    )}
+                    {bangWeetObj.attachmentUrl && <img src={bangWeetObj.attachmentUrl}/>}
                     {isOwner && (
-                        <>
-                            <button onClick={toggleEditting}>Edit bangWeet</button>
-                            <button onClick={onDeleteClick}>Delete bangWeet</button>
-                        </>
+                        <div className="bangWeet__actions">
+                            <span onClick={onDeleteClick}>
+                                <FontAwesomeIcon icon={faTrash}/>
+                            </span>
+                            <span onClick={toggleEditting}>
+                                <FontAwesomeIcon icon={faPencilAlt}/>
+                            </span>
+                        </div>
                     )}
                 </>
             }
